fix(glm_moe_dsa): cast router logits to float32 to match HF base class
GlmMoeDsaTopkRouter.forward() in HuggingFace explicitly casts both
hidden_states and weight to float32 before F.linear.  QEffGlmMoeDsaTopkRouter
was computing logits in the activation dtype (float32 in the current test
but could differ in production) which causes slightly different routing
scores and therefore different expert selections.

This change makes the router numerically consistent with HF regardless of
the activation dtype of upstream tensors.

Note: HF ≠ QEff token divergence is not fully resolved by this fix alone —
the QEff model applies Dynamic Sparse Attention (DSA) via the indexer which
creates a sparse attention mask, while the HF baseline uses full causal
attention.  These are architecturally different computations; exact token
match with HF requires running in full-attention mode.

+++ QEfficient/transformers/models/glm_moe_dsa/modeling_glm_moe_dsa.py
@@ -548,7 +548,11 @@ class QEffGlmMoeDsaSparseDecoderLayer(QEffGlmMoeDsaDecoderLayer):
 class QEffGlmMoeDsaTopkRouter(GlmMoeDsaTopkRouter):
     def forward(self, hidden_states):
         hidden_states = hidden_states.view(-1, self.config.hidden_size)
-        router_logits = F.linear(hidden_states, self.weight)
+        # Cast to float32 to match HF base class (GlmMoeDsaTopkRouter.forward uses
+        # hidden_states.type(float32) and self.weight.type(float32)).  Without this,
+        # router logits are computed in the activation dtype and small numerical
+        # differences change which experts are selected, causing HF ≠ QEff divergence.
+        router_logits = F.linear(hidden_states.to(torch.float32), self.weight.to(torch.float32))
         router_scores = router_logits.sigmoid()
         scores_for_choice = router_scores + self.e_score_correction_bias.unsqueeze(0).to(router_scores.device)
         group_scores = torch.einsum(
